fix: move slices to folder store
