+++ src/store/slices/theme.slice.ts
